Fix datetime\diff returning float for days, hours and minutes
floor() returns a float, so datetime\diff with 'days', 'hours' or
'minutes' produced 3.0 instead of 3. Formula comparison is strict, which
made an expression like datetime\diff(a, b, 'days') == 3 evaluate to
false. Cast the result to int, as the function is declared to return.

+++ application/Espo/Core/Formula/Functions/DatetimeGroup/DiffType.php
@@ -119,11 +119,11 @@ class DiffType extends BaseFunction
         if ($intervalType === 'seconds') {
             $number = $secondsDiff;
         } else if ($intervalType === 'minutes') {
-            $number = floor($secondsDiff / 60);
+            $number = (int) floor($secondsDiff / 60);
         } else if ($intervalType === 'hours') {
-            $number = floor($secondsDiff / (60 * 60));
+            $number = (int) floor($secondsDiff / (60 * 60));
         } else if ($intervalType === 'days') {
-            $number = floor($secondsDiff / (60 * 60 * 24));
+            $number = (int) floor($secondsDiff / (60 * 60 * 24));
         } else {
             $property = $this->intervalTypePropertyMap[$intervalType];
             $interval = $dateTime2->diff($dateTime1);
